Rename language to track_id
Rename the `language` column in the user_exercises and submissions
tables to `track_id`, since 'language' had been a misnomer for what was
actually a track_id.

Add getter and setter methods that substitute #track_id for #language.
Eventually, when all references to `language` are all updated to be
references to `track_id` in the main code, then these temporary methods
will be able to be removed. The temporary methods are in:
`lib/exercism/temporary_language_to_track_id_translator.rb`

Do a *bare minimum* mass renaming of `language` to `track_id`. Bare
minimum means just enough to get all the tests back to passing.
Eventually, we'll go through the codebase in more detail and rename all
the affected places so that the temporary methods mentioned above are no
longer necessary.